refact: test tool
