+++ tests/Intelligence/Tools/GuardsRepeatCallsTest.php
@@ -6,10 +6,16 @@ namespace Tests\Intelligence\Tools;
 
 use Kanvas\Intelligence\Agents\Enums\ToolOutcomeEnum;
 use Kanvas\Intelligence\Agents\Neuron\Tools\Traits\GuardsRepeatCalls;
+use ReflectionClass;
+use ReflectionProperty;
+use SplFileInfo;
+use Symfony\Component\Finder\Finder;
 use Tests\TestCase;
 
 class GuardsRepeatCallsTest extends TestCase
 {
+    private const string TOOLS_PATH = 'src/Domains/Intelligence/Agents/Neuron/Tools';
+
     public function test_the_second_identical_call_does_not_execute_the_work(): void
     {
         $tool = new RepeatGuardedProbe();
@@ -64,6 +70,73 @@ class GuardsRepeatCallsTest extends TestCase
 
         $this->assertSame(1, $tool->executions);
     }
+
+    /**
+     * The whole guard hangs on this. NeuronAI never invokes the registered tool — it hands each call a
+     * shallow `clone` (`HandleWithTools::findTool`), which shares the ledger OBJECT but would copy a
+     * null. A ledger built lazily is therefore built on the clone, every call gets its own empty one,
+     * and the guard silently never fires. That is how it shipped, unused, before KANVAS-ECOSYSTEM-6A1.
+     */
+    public function test_the_guard_still_holds_across_the_clones_neuron_makes_per_call(): void
+    {
+        $registered = new RepeatGuardedProbe();
+
+        $firstCall = clone $registered;
+        $secondCall = clone $registered;
+
+        $firstCall->run(['name' => 'acme']);
+        $second = $secondCall->run(['name' => 'acme']);
+
+        $this->assertArrayHasKey('repeat_call', $second, 'The ledger did not survive the clone.');
+        $this->assertSame(0, $secondCall->executions, 'The second clone re-ran the work.');
+    }
+
+    public function test_every_guarded_tool_initialises_its_ledger_in_the_constructor(): void
+    {
+        $guarded = $this->guardedTools();
+
+        $this->assertNotEmpty($guarded, 'No tool uses GuardsRepeatCalls — did the trait lose its last caller?');
+
+        foreach ($guarded as $class) {
+            $constructor = new ReflectionClass($class)->getConstructor();
+
+            $this->assertNotNull($constructor, $class . ' uses GuardsRepeatCalls but has no constructor to init it.');
+            $this->assertSame(
+                0,
+                $constructor->getNumberOfRequiredParameters(),
+                $class . ' takes constructor arguments; add it to this test explicitly so the guard stays covered.',
+            );
+
+            $tool = new $class();
+
+            $this->assertTrue(
+                new ReflectionProperty($tool, 'repeatLedger')->isInitialized($tool),
+                $class . ' must call initRepeatGuard() in its constructor — see '
+                    . 'test_the_guard_still_holds_across_the_clones_neuron_makes_per_call for why it cannot be lazy.',
+            );
+        }
+    }
+
+    /**
+     * @return array<int, class-string>
+     */
+    private function guardedTools(): array
+    {
+        $files = new Finder()
+            ->files()
+            ->in(base_path(self::TOOLS_PATH))
+            ->name('*.php')
+            ->notPath('Traits')
+            ->contains('use GuardsRepeatCalls;');
+
+        return array_values(array_map(
+            static fn (SplFileInfo $file): string => 'Kanvas\\Intelligence\\Agents\\Neuron\\Tools\\'
+                . str_replace('/', '\\', $file->getRelativePath())
+                . ($file->getRelativePath() === '' ? '' : '\\')
+                . $file->getBasename('.php'),
+            iterator_to_array($files),
+        ));
+    }
 }
 
 class RepeatGuardedProbe
@@ -71,6 +144,15 @@ class RepeatGuardedProbe
     use GuardsRepeatCalls;
 
     public int $executions = 0;
+
+    /**
+     * Models the real contract: a guarded tool builds its ledger in the constructor, because that is
+     * the only point that runs before NeuronAI starts cloning it per call.
+     */
+    public function __construct()
+    {
+        $this->initRepeatGuard();
+    }
 
     /**
      * @param array<string, mixed> $inputs
